Experimental fix on low memory device
Upstream ffmpeg.wasm use INITIAL_MEMORY of 1GB, this add another custom
build with INITIAL_MEMORY of 500MB. The browser will decide which version
to use is based on total device memory.

`navigator.deviceMemory` is not available on Firefox/Firefox Mobile
so it will default to 1GB version. I will ignore this for now since this
fix is meant for Chrome and 1GB version doesn't have issue on other
browser anyway.

Another experiment is to call `ffmpeg.exit()` in hope of avoiding
memory leak?